Split the helper module to its own component
