+++ src/analysis/retail/evoker/augmentation/modules/breahtofeons/BreathOfEonsRotational.tsx
@@ -29,6 +29,7 @@ import trinkets from 'common/ITEMS/dragonflight/trinkets';
 import Combatant from 'parser/core/Combatant';
 import Combatants from 'parser/shared/modules/Combatants';
 import { SpellTracker } from 'analysis/retail/evoker/shared/modules/components/ExplanationGraph';
+import BreathOfEonsHelper from './BreathOfEonsHelper';
 
 export type BreathOfEonsWindows = {
   flightData: SpellTracker[];
@@ -574,6 +575,18 @@ class BreathOfEonsRotational extends Analyzer {
         fightEndTime={this.owner.fight.end_time}
         ebonMightCount={this.ebonMightCount}
         shiftingSandsCount={this.shiftingSandsCount}
+      />
+    );
+  }
+  helperSection(): JSX.Element | null {
+    if (!this.active) {
+      return null;
+    }
+    return (
+      <BreathOfEonsHelper
+        windows={this.windows}
+        fightStartTime={this.owner.fight.start_time}
+        fightEndTime={this.owner.fight.end_time}
         owner={this.owner}
       />
     );
